# gui/src/assets/openapi.json
{
    "openapi": "3.1.0",
    "info": {
    "title": "FastAPI",
    "version": "0.1.0"
    },
    "paths": {
...
    "/api/v1/med-records/": {
    "post": {
    "tags": [
    "MedRecords"
    ],
    "summary": "Medrecordcontroller.Create Record",
    "description": "Create a new medical record.",
    "operationId": "MedRecordController_create_record_api_v1_med_records__post",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/MedRecordCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/MedRecordRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "MedRecords"
    ],
    "summary": "Medrecordcontroller.List Records",
    "operationId": "MedRecordController_list_records_api_v1_med_records__get",
    "parameters": [
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    },
    {
    "name": "patient_id",
    "in": "query",
    "required": false,
    "schema": {
    "anyOf": [
    {
    "type": "string",
    "format": "uuid"
    },
    {
    "type": "null"
    }
    ],
    "title": "Patient Id"
    }
    },
    {
    "name": "doctor_id",
    "in": "query",
    "required": false,
    "schema": {
    "anyOf": [
    {
    "type": "string",
    "format": "uuid"
    },
    {
    "type": "null"
    }
    ],
    "title": "Doctor Id"
    }
    },
    {
    "name": "department_id",
    "in": "query",
    "required": false,
    "schema": {
    "anyOf": [
    {
    "type": "integer"
    },
    {
    "type": "null"
    }
    ],
    "title": "Department Id"
    }
    },
    {
    "name": "examination_date",
    "in": "query",
    "required": false,
    "schema": {
    "anyOf": [
    {
    "type": "string",
    "format": "date-time"
    },
    {
    "type": "null"
    }
    ],
    "title": "Examination Date"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_MedRecordRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/med-records/{record_id}/transcriptions": {
    "post": {
    "tags": [
    "MedRecords"
    ],
    "summary": "Medrecordcontroller.Create Transcription",
    "operationId": "MedRecordController_create_transcription_api_v1_med_records__record_id__transcriptions_post",
    "parameters": [
    {
    "name": "record_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/TranscriptionCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/TranscriptionRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "MedRecords"
    ],
    "summary": "Medrecordcontroller.List Transcriptions",
    "operationId": "MedRecordController_list_transcriptions_api_v1_med_records__record_id__transcriptions_get",
    "parameters": [
    {
    "name": "record_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    }
    },
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_TranscriptionRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/med-records/{record_id}": {
    "delete": {
    "tags": [
    "MedRecords"
    ],
    "summary": "Medrecordcontroller.Delete Record",
    "description": "Delete a medical record.",
    "operationId": "MedRecordController_delete_record_api_v1_med_records__record_id__delete",
    "parameters": [
    {
    "name": "record_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "MedRecords"
    ],
    "summary": "Medrecordcontroller.Get Record",
    "description": "Get a medical record by ID.",
    "operationId": "MedRecordController_get_record_api_v1_med_records__record_id__get",
    "parameters": [
    {
    "name": "record_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/MedRecordDetailRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "patch": {
    "tags": [
    "MedRecords"
    ],
    "summary": "Medrecordcontroller.Update Record",
    "description": "Update a medical record.",
    "operationId": "MedRecordController_update_record_api_v1_med_records__record_id__patch",
    "parameters": [
    {
    "name": "record_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/MedRecordUpdate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/MedRecordRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/med-records/{record_id}/transcriptions/{transcription_id}": {
    "delete": {
    "tags": [
    "MedRecords"
    ],
    "summary": "Medrecordcontroller.Delete Transcription",
    "operationId": "MedRecordController_delete_transcription_api_v1_med_records__record_id__transcriptions__transcription_id__delete",
    "parameters": [
    {
    "name": "record_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    }
    },
    {
    "name": "transcription_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Transcription Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "MedRecords"
    ],
    "summary": "Medrecordcontroller.Get Transcription",
    "operationId": "MedRecordController_get_transcription_api_v1_med_records__record_id__transcriptions__transcription_id__get",
    "parameters": [
    {
    "name": "record_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    }
    },
    {
    "name": "transcription_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Transcription Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/TranscriptionDetailRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    },
    "components": {
    "schemas": {
...
    "DepartmentRead": {
    "properties": {
    "department_id": {
    "type": "integer",
    "title": "Department Id"
    },
    "name": {
    "type": "string",
    "title": "Name"
    },
    "description": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Description"
    },
    "contact_phone": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Contact Phone"
    }
    },
    "type": "object",
    "required": [
    "department_id",
    "name"
    ],
    "title": "DepartmentRead"
    },
...
    "DoctorRead": {
    "properties": {
    "doctor_id": {
    "type": "string",
    "format": "uuid",
    "title": "Doctor Id"
    },
    "first_name": {
    "type": "string",
    "title": "First Name"
    },
    "last_name": {
    "type": "string",
    "title": "Last Name"
    },
    "middle_name": {
    "type": "string",
    "title": "Middle Name"
    },
    "email": {
    "type": "string",
    "title": "Email"
    },
    "phone_number": {
    "type": "string",
    "title": "Phone Number"
    },
    "birth_data": {
    "type": "string",
    "format": "date",
    "title": "Birth Data"
    },
    "is_active": {
    "type": "boolean",
    "title": "Is Active"
    },
    "gender": {
    "$ref": "#/components/schemas/Gender"
    },
    "specialization": {
    "type": "string",
    "title": "Specialization"
    },
    "department": {
    "$ref": "#/components/schemas/DepartmentRead"
    }
    },
    "type": "object",
    "required": [
    "doctor_id",
    "first_name",
    "last_name",
    "middle_name",
    "email",
    "phone_number",
    "birth_data",
    "is_active",
    "gender",
    "specialization",
    "department"
    ],
    "title": "DoctorRead"
    },
...
    "Gender": {
    "type": "string",
    "enum": [
    "male",
    "female"
    ],
    "title": "Gender"
    },
    "HTTPValidationError": {
    "properties": {
    "detail": {
    "items": {
    "$ref": "#/components/schemas/ValidationError"
    },
    "type": "array",
    "title": "Detail"
    }
    },
    "type": "object",
    "title": "HTTPValidationError"
    },
...
    "MedRecordCreate": {
    "properties": {
    "patient_id": {
    "type": "string",
    "format": "uuid",
    "title": "Patient Id"
    },
    "doctor_id": {
    "type": "string",
    "format": "uuid",
    "title": "Doctor Id"
    },
    "department_id": {
    "type": "integer",
    "title": "Department Id"
    },
    "examination_date": {
    "type": "string",
    "format": "date-time",
    "title": "Examination Date"
    }
    },
    "type": "object",
    "required": [
    "patient_id",
    "doctor_id",
    "department_id",
    "examination_date"
    ],
    "title": "MedRecordCreate"
    },
    "MedRecordDetailRead": {
    "properties": {
    "patient_id": {
    "type": "string",
    "format": "uuid",
    "title": "Patient Id"
    },
    "doctor_id": {
    "type": "string",
    "format": "uuid",
    "title": "Doctor Id"
    },
    "department_id": {
    "type": "integer",
    "title": "Department Id"
    },
    "examination_date": {
    "type": "string",
    "format": "date-time",
    "title": "Examination Date"
    },
    "record_id": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    },
    "patient": {
    "$ref": "#/components/schemas/PatientRead"
    },
    "doctor": {
    "$ref": "#/components/schemas/DoctorRead"
    },
    "department": {
    "$ref": "#/components/schemas/DepartmentRead"
    },
    "transcriptions": {
    "items": {
    "$ref": "#/components/schemas/TranscriptionRead"
    },
    "type": "array",
    "title": "Transcriptions",
    "default": []
    }
    },
    "type": "object",
    "required": [
    "patient_id",
    "doctor_id",
    "department_id",
    "examination_date",
    "record_id",
    "patient",
    "doctor",
    "department"
    ],
    "title": "MedRecordDetailRead"
    },
...
    "MedRecordRead": {
    "properties": {
    "patient_id": {
    "type": "string",
    "format": "uuid",
    "title": "Patient Id"
    },
    "doctor_id": {
    "type": "string",
    "format": "uuid",
    "title": "Doctor Id"
    },
    "department_id": {
    "type": "integer",
    "title": "Department Id"
    },
    "examination_date": {
    "type": "string",
    "format": "date-time",
    "title": "Examination Date"
    },
    "record_id": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    },
    "patient": {
    "$ref": "#/components/schemas/PatientRead"
    },
    "doctor": {
    "$ref": "#/components/schemas/DoctorRead"
    },
    "department": {
    "$ref": "#/components/schemas/DepartmentRead"
    }
    },
    "type": "object",
    "required": [
    "patient_id",
    "doctor_id",
    "department_id",
    "examination_date",
    "record_id",
    "patient",
    "doctor",
    "department"
    ],
    "title": "MedRecordRead"
    },
    "MedRecordUpdate": {
    "properties": {
    "patient_id": {
    "anyOf": [
    {
    "type": "string",
    "format": "uuid"
    },
    {
    "type": "null"
    }
    ],
    "title": "Patient Id"
    },
    "department_id": {
    "anyOf": [
    {
    "type": "integer"
    },
    {
    "type": "null"
    }
    ],
    "title": "Department Id"
    },
    "examination_date": {
    "anyOf": [
    {
    "type": "string",
    "format": "date-time"
    },
    {
    "type": "null"
    }
    ],
    "title": "Examination Date"
    }
    },
    "type": "object",
    "title": "MedRecordUpdate"
    },
...
    "Page_MedRecordRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/MedRecordRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "total",
    "page",
    "size"
    ],
    "title": "Page[MedRecordRead]"
    },
...
    "Page_TranscriptionRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/TranscriptionRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "total",
    "page",
    "size"
    ],
    "title": "Page[TranscriptionRead]"
    },
...
    "PatientRead": {
    "properties": {
    "patient_id": {
    "type": "string",
    "format": "uuid",
    "title": "Patient Id"
    },
    "first_name": {
    "type": "string",
    "title": "First Name"
    },
    "last_name": {
    "type": "string",
    "title": "Last Name"
    },
    "middle_name": {
    "type": "string",
    "title": "Middle Name"
    },
    "phone_number": {
    "type": "string",
    "title": "Phone Number"
    },
    "email": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Email"
    },
    "birth_data": {
    "type": "string",
    "format": "date",
    "title": "Birth Data"
    },
    "is_active": {
    "type": "boolean",
    "title": "Is Active"
    }
    },
    "type": "object",
    "required": [
    "patient_id",
    "first_name",
    "last_name",
    "middle_name",
    "phone_number",
    "email",
    "birth_data",
    "is_active"
    ],
    "title": "PatientRead"
    },
...
    "TranscriptionCreate": {
    "properties": {
    "record_id": {
    "anyOf": [
    {
    "type": "string",
    "format": "uuid"
    },
    {
    "type": "null"
    }
    ],
    "title": "Record Id"
    },
    "audio_url": {
    "type": "string",
    "title": "Audio Url"
    },
    "audio_duration": {
    "type": "integer",
    "title": "Audio Duration",
    "default": 0
    },
    "transcription_text": {
    "type": "string",
    "title": "Transcription Text"
    },
    "asr_model": {
    "type": "string",
    "title": "Asr Model",
    "default": "whisper"
    },
    "asr_config": {
    "type": "object",
    "title": "Asr Config",
    "default": {}
    },
    "language": {
    "type": "string",
    "title": "Language",
    "default": "uk"
    }
    },
    "type": "object",
    "required": [
    "audio_url",
    "transcription_text"
    ],
    "title": "TranscriptionCreate"
    },
    "TranscriptionDetailRead": {
    "properties": {
    "record_id": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    },
    "transcription_id": {
    "type": "integer",
    "title": "Transcription Id"
    },
    "audio_url": {
    "type": "string",
    "title": "Audio Url"
    },
    "audio_duration": {
    "type": "integer",
    "title": "Audio Duration"
    },
    "transcription_text": {
    "type": "string",
    "title": "Transcription Text"
    },
    "asr_model": {
    "type": "string",
    "title": "Asr Model"
    },
    "asr_config": {
    "type": "object",
    "title": "Asr Config"
    },
    "language": {
    "type": "string",
    "title": "Language"
    }
    },
    "type": "object",
    "required": [
    "record_id",
    "transcription_id",
    "audio_url",
    "audio_duration",
    "transcription_text",
    "asr_model",
    "asr_config",
    "language"
    ],
    "title": "TranscriptionDetailRead"
    },
    "TranscriptionRead": {
    "properties": {
    "record_id": {
    "type": "string",
    "format": "uuid",
    "title": "Record Id"
    },
    "transcription_id": {
    "type": "integer",
    "title": "Transcription Id"
    },
    "audio_url": {
    "type": "string",
    "title": "Audio Url"
    },
    "audio_duration": {
    "type": "integer",
    "title": "Audio Duration"
    },
    "transcription_text": {
    "type": "string",
    "title": "Transcription Text"
    }
    },
    "type": "object",
    "required": [
    "record_id",
    "transcription_id",
    "audio_url",
    "audio_duration",
    "transcription_text"
    ],
    "title": "TranscriptionRead"
    },
    "ValidationError": {
    "properties": {
    "loc": {
    "items": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "integer"
    }
    ]
    },
    "type": "array",
    "title": "Location"
    },
    "msg": {
    "type": "string",
    "title": "Message"
    },
    "type": {
    "type": "string",
    "title": "Error Type"
    }
    },
    "type": "object",
    "required": [
    "loc",
    "msg",
    "type"
    ],
    "title": "ValidationError"
    }
    },
    "securitySchemes": {
    "APIKeyCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "med_voice_auth_token"
    }
    }
    }
    }